[+] - Check Firmware Both side (OK) | Now fix some live issue (ble gatt & AT+CWAP)
