+++ com-at/main/com-at.c
@@ -1,9 +1,12 @@
 // Eun0us - DVID - main.c
 
+#include <string.h>
 #include <stdio.h>
 #include "esp_log.h"
 #include "nvs_flash.h"
 #include "driver/uart.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
 
 #include "at_server.h"
 #include "at_wifi.h"  // Ton module AT Wi-Fi
@@ -11,18 +14,28 @@
 #include "at_ota.h"
 //#include "at_mqtt.h"
 
+
+#define BUF_SIZE 1024
+
+// Mets ici les pins que tu veux tester ! IO1/IO0 etc.
+#define UART_TX_PIN         1      // IO1
+#define UART_RX_PIN         0      // IO0
+
 // Initialisation de l'UART pour AT et debug            
 void uart_init(void)
 {
-    const uart_config_t uart_config = {
+    uart_config_t uart_config = {
         .baud_rate = 115200,
         .data_bits = UART_DATA_8_BITS,
         .parity    = UART_PARITY_DISABLE,
         .stop_bits = UART_STOP_BITS_1,
-        .flow_ctrl = UART_HW_FLOWCTRL_DISABLE
+        .flow_ctrl = UART_HW_FLOWCTRL_DISABLE,
+        .source_clk = UART_SCLK_DEFAULT,
     };
-    uart_param_config(UART_NUM_0, &uart_config);
-    uart_driver_install(UART_NUM_0, 2048, 0, 0, NULL, 0);
+
+    uart_driver_install(UART_NUM, BUF_SIZE, 0, 0, NULL, 0);
+    uart_param_config(UART_NUM, &uart_config);
+    uart_set_pin(UART_NUM, UART_TX_PIN, UART_RX_PIN, UART_PIN_NO_CHANGE, UART_PIN_NO_CHANGE);
 }
 
 void app_main(void)
