+++ test/main/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "test.c"
+                    INCLUDE_DIRS ".")
